social_pulse/social-pulse-react-app/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
mysqlclient>=2.2
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
lightgbm>=4.1
catboost>=1.2
ydata-profiling>=4.6
joblib>=1.3
dj-database-url
gunicorn
whitenoise